nixd: refactor files of nixd implementation
Split nixd implementation into multiple files to make it more readable.
